moved pages into cpp
